requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
sqlite3
mysql-connector